Improvement of the combobox widget functionality
I was finalizing the functionality of the combobox widget. The program loads records depending on the group, but has not yet finished showing the groups depending on the user. There is also a bug on the last group. It is better not to choose it, because it breaks the program. You need to make a condition for the request (if nothing is found by the group ID, then you need to load empty tables